SPARKC-348 accept predicates with indexed partition columns
Preconditions with indexed columns that belong to partition key but do
not form whole partition key generate failures.
This kind of queries should be executed as a query with index column
specified, not discarded as a query with incomplete partition key.

This changes CassandraRDDPartitioner#containsPartitionKey so that it
does not throw exception whenever the partition key is incomplete and
all precodition colums are indexed. Other cases for
CassandraRDDPartitioner#containsPartitionKey are not changed.